Be more aggressive about repainting page overlays https://bugs.webkit.org/show_bug.cgi?id=80336 <rdar://problem/10965943>
Reviewed by Simon Fraser.

Source/WebCore:

Add a way to find out if a given GraphicsLayer is going to be repainted.

* WebCore.exp.in:
* platform/graphics/GraphicsLayer.h:
(WebCore::GraphicsLayer::needsDisplay):
(GraphicsLayer):
* platform/graphics/ca/GraphicsLayerCA.h:

Source/WebKit2:

Whenever we're flushing layers and we have a page overlay, check if the main frame
has scrolled or if the main frame root content layer needs to be repainted and force the
overlay layer to be repainted if either of those conditions are true.

* WebProcess/WebPage/mac/TiledCoreAnimationDrawingArea.h:
(TiledCoreAnimationDrawingArea):
* WebProcess/WebPage/mac/TiledCoreAnimationDrawingArea.mm:
(WebKit::TiledCoreAnimationDrawingArea::flushLayers):
(WebKit::TiledCoreAnimationDrawingArea::shouldRepaintPageOverlayLayer):
(WebKit):

Identifier: 97461@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@109812 268f45cc-cd09-0410-ab3c-d52691b4dbfc