Fix DerivedSources.make to not assume that WTF and WebCore build in to the same build directory.
We can't assume that wtf/Platform.h can be found inside BUILT_PRODUCTS_DIR as
that won't be the case if WTF and WebCore are building in to different build
directories. We should instead look for wtf/Platform.h in the header search path.

Unreviewed due to being a build fix.

* DerivedSources.make: Set up the header search path, and invoke the compiler in
such a manner that it looks for wtf/Platform.h in that search path.

Identifier: 97659@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@110042 268f45cc-cd09-0410-ab3c-d52691b4dbfc