Do not refer to resutlsButtonElement and cancelButtonElement to compute paddings of search popups https://bugs.webkit.org/show_bug.cgi?id=80564
Reviewed by Hajime Morita.

We'd like to reduce dependency of resultsButtonElement and
cancelButtonElement from RenderTextControlSingleLine.

No behavior change.

* rendering/RenderTextControlSingleLine.cpp:
(WebCore::RenderTextControlSingleLine::clientPaddingLeft):
Use the left position of innerBlockElement, instead of the width of resultsButtonElement.
They are equivalent.
(WebCore::RenderTextControlSingleLine::clientPaddingRight):
Use <the container width - right position of innerBlockElement>,
instead of the width of the cancelButtonElement. They are
equivalent.

Identifier: 97751@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@110145 268f45cc-cd09-0410-ab3c-d52691b4dbfc