Crash in SVGTextLayoutAttributesBuilder::fillCharacterDataMap https://bugs.webkit.org/show_bug.cgi?id=78949 <rdar://problem/10889440>
Reviewed by Nikolas Zimmermann.

Invalidate the text positioning cache when the children of an SVGAElement change,
so that we regenerate the list the next time it's needed instead of using stale values.

Test: svg/text/text-positioning-remove-child-crash.svg

* rendering/svg/SVGAElement.cpp:
(WebCore::SVGAElement::childrenChanged):

Add a test ensuring that we don't crash when removing a child of <a> inside a <text> block.

* svg/text/text-positioning-remove-child-crash-expected.txt: Added.
* svg/text/text-positioning-remove-child-crash.svg: Added.


Identifier: 97903@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@110307 268f45cc-cd09-0410-ab3c-d52691b4dbfc