<rdar://problem/11025225> Assertion failure in RenderView::computeRectForRepaint() (!repaintContainer || repaintContainer == this) at store.apple.com https://bugs.webkit.org/show_bug.cgi?id=81051
Reviewed by Simon Fraser.

.:

* ManualTests/inline-repaint-container.html: Added.

Source/WebCore:

Test: ManualTests/inline-repaint-container.html.

* rendering/RenderInline.cpp:
(WebCore::RenderInline::clippedOverflowRectForRepaint): This function was not handling the
case of the repaint container being a descendant of the containing block correctly, leading
to the assertion failure, but also to a correctness bug seen in the new test. If the repaint
container is a descendant of the containing block, just return the rect in the repaint
container coordinates.



Identifier: 98219@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@110643 268f45cc-cd09-0410-ab3c-d52691b4dbfc