[Textmap] Disable driver based BGRA swizzling for OpenGL/ES
Reviewed by Noam Rosenthal.

The current way of detecting swizzling support (check for GL_EXT_texture_format_BGRA8888)
is unfortunately not reliable. We need a better way of detection that (filed bug 81103).
In the meantime this patch reverts back to doing the swizzling in the texture mapper to
get pixels back onto the screen.

* platform/graphics/texmap/TextureMapperGL.cpp:
(WebCore::driverSupportsBGRASwizzling):
(WebCore::BitmapTextureGL::updateContents):

Identifier: 98277@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@110703 268f45cc-cd09-0410-ab3c-d52691b4dbfc