java/java-and-plugins.html test failing https://bugs.webkit.org/show_bug.cgi?id=81188 <rdar://problem/11048657>
Reviewed by Sam Weinig.

Source/WebKit/mac:

When we're marshaling JSObjects that wrap plug-in objects, make sure to check that the
current plug-in instance is the same as the plug-in instance the object came from and don't
pass the unwrapped object ID if that is the case.

* Plugins/Hosted/NetscapePluginInstanceProxy.mm:
(WebKit::getObjectID):
(WebKit::NetscapePluginInstanceProxy::addValueToArray):
(WebKit::NetscapePluginInstanceProxy::marshalValues):
* Plugins/Hosted/ProxyInstance.h:
(ProxyInstance):
(WebKit::ProxyInstance::instanceProxy):

Source/WebKit2:

When we're marshaling NPObjects that wrap plug-in objects, make sure to check that the
current plug-in instance is the same as the plug-in instance the object came from and don't
pass the unwrapped object ID if that is the case.

* Shared/Plugins/NPRemoteObjectMap.cpp:
(WebKit::remoteNPObjectID):
(WebKit::NPRemoteObjectMap::npVariantToNPVariantData):

LayoutTests:

Remove part of this test that tested trying to call a java method on an NPObject; on Snow Leopard
and newer the Java plug-in is also a NPAPI plug-in so the two removed tests would fail.

* java/java-and-plugins-expected.txt:
* java/java-and-plugins.html:

Identifier: 98382@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@110820 268f45cc-cd09-0410-ab3c-d52691b4dbfc