Assertion failures in WebCore::Page::goBackOrForward causing multiple "crashes" on Lion Intel Debug WebKit2 testers https://bugs.webkit.org/show_bug.cgi?id=81159
Rubber-stamped by Brady Eidson.

It is plausible that history could change in between the time a navigation request is
scheduled and the timer is fired, so the request index could be out of the b/f list range.
Hence, the assertion is invalid and should be removed.

* page/Page.cpp:
(WebCore::Page::goBackOrForward):

Identifier: 98446@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@110884 268f45cc-cd09-0410-ab3c-d52691b4dbfc