Simplify SmallStrings https://bugs.webkit.org/show_bug.cgi?id=81445
Patch by Benjamin Poulain <bpoulain@apple.com> on 2012-03-19
Reviewed by Gavin Barraclough.

SmallStrings had two methods that should not be public: count() and clear().

The method clear() is effectively replaced by finalizeSmallStrings(). The body
of the method was moved to the constructor since the code is obvious.

The method count() is unused.

* runtime/SmallStrings.cpp:
(JSC::SmallStrings::SmallStrings):
* runtime/SmallStrings.h:
(SmallStrings):


Identifier: 98836@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@111306 268f45cc-cd09-0410-ab3c-d52691b4dbfc