[Qt][WK2][Mac] MiniBrowser asserts at HashTable.h:480 https://bugs.webkit.org/show_bug.cgi?id=81113
The GLContext is created by QWindow/QSurface using
QOpenGLContext which is implemented using NSOpenGL functions
on mac. Therefore we should use QOpenGLContext as well when
querying for the current context. AGL is not supposed
to be paired with NSOpenGL.
Altering the typedef for GLContext is fine, since
the GLContext type is only used as a key in a map.

Reviewed by Kenneth Rohde Christiansen.

* platform/graphics/texmap/TextureMapperGL.cpp:
(SharedGLData):
(WebCore::TextureMapperGLData::SharedGLData::getCurrentGLContext):

Identifier: 99027@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@111524 268f45cc-cd09-0410-ab3c-d52691b4dbfc